Adjusting some features and finalizing order features

+++ src/main/java/com/mdp/dto/request/ProductRequestDTO.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ProductRequestDTO(
         Long id,
 
@@ -15,11 +17,12 @@ public record ProductRequestDTO(
 
         @NotNull(message = "Price is required")
         @Positive(message = "Price must be positive")
-        Double price,
+        BigDecimal price,
 
         @NotNull(message = "Stock is required")
         @Min(value = 0, message = "Stock cannot be negative")
         Integer stock
+
 ) {}
