+++ src/main/java/com/mdp/dto/response/ProductResponseDTO.java
@@ -1,10 +1,12 @@
 package com.mdp.dto.response;
 
+import java.math.BigDecimal;
+
 public record ProductResponseDTO(
         Long id,
         String name,
         String description,
-        Double price,
+        BigDecimal price,
         Integer stock
 ) {}
